refactored file placement
