+++ src/main/scala/ai/privado/languageEngine/kotlin/tagger/PrivadoTagger.scala
@@ -2,16 +2,16 @@ package ai.privado.languageEngine.kotlin.tagger
 
 import ai.privado.cache.{DataFlowCache, RuleCache, TaggerCache}
 import ai.privado.entrypoint.PrivadoInput
+import ai.privado.feeder.PermissionSourceRule
 import ai.privado.languageEngine.java.feeder.StorageInheritRule
 import ai.privado.languageEngine.java.tagger.config.JavaDBConfigTagger
 import ai.privado.languageEngine.java.tagger.sink.{InheritMethodTagger, JavaAPITagger}
 import ai.privado.languageEngine.java.tagger.source.{IdentifierTagger, InSensitiveCallTagger}
-import ai.privado.languageEngine.kotlin.feeder.{PermissionSourceRule, StorageAnnotationRule}
+import ai.privado.languageEngine.kotlin.feeder.StorageAnnotationRule
 import ai.privado.languageEngine.kotlin.tagger.sink.StorageAnnotationTagger
-import ai.privado.languageEngine.kotlin.tagger.source.AndroidXMLPermissionTagger
 import ai.privado.tagger.PrivadoBaseTagger
 import ai.privado.tagger.sink.{APITagger, RegularSinkTagger}
-import ai.privado.tagger.source.{LiteralTagger, SqlQueryTagger}
+import ai.privado.tagger.source.{AndroidXmlPermissionTagger, LiteralTagger, SqlQueryTagger}
 import io.shiftleft.codepropertygraph.generated.Cpg
 import io.shiftleft.codepropertygraph.generated.nodes.Tag
 import io.shiftleft.semanticcpg.language.*
@@ -37,7 +37,7 @@ class PrivadoTagger(cpg: Cpg) extends PrivadoBaseTagger {
 
     new InSensitiveCallTagger(cpg, ruleCache, taggerCache).createAndApply()
 
-    new AndroidXMLPermissionTagger(cpg, ruleCache, PermissionSourceRule.miniatureRuleList).createAndApply()
+    new AndroidXmlPermissionTagger(cpg, ruleCache, PermissionSourceRule.miniatureRuleList).createAndApply()
 
     new JavaDBConfigTagger(cpg).createAndApply()
